move classes to the correct package
