+++ app/src/main/java/com/android/dev/engineer/kotlin/compose/data/domain/local/MainNavGraph.kt
@@ -1,4 +1,4 @@
-package com.android.dev.engineer.kotlin.compose.data.domain
+package com.android.dev.engineer.kotlin.compose.data.domain.local
 
 sealed class MainNavGraph(val route: String) {
     object Intro : MainNavGraph(route = "intro")
