+++ app/src/main/java/com/android/dev/engineer/kotlin/compose/feature/routing/RoutingViewModel.kt
@@ -4,7 +4,7 @@ package com.android.dev.engineer.kotlin.compose.feature.routing
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.android.dev.engineer.kotlin.compose.data.domain.MainNavGraph
+import com.android.dev.engineer.kotlin.compose.data.domain.local.MainNavGraph
 import com.android.dev.engineer.kotlin.compose.data.use_case.initial_route.GetInitialRouteUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
