+++ app/src/test/java/com/android/dev/engineer/kotlin/compose/test/data/use_case/GetInitialRouteUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.android.dev.engineer.kotlin.compose.test.data.use_case
 
-import com.android.dev.engineer.kotlin.compose.data.domain.MainNavGraph
+import com.android.dev.engineer.kotlin.compose.data.domain.local.MainNavGraph
 import com.android.dev.engineer.kotlin.compose.data.use_case.initial_route.GetInitialRouteUseCase
 import com.android.dev.engineer.kotlin.compose.data.use_case.initial_route.GetInitialRouteUseCaseImpl
 import com.android.dev.engineer.kotlin.compose.fake.AppDataStoreFake
